refactor(manager): fix some typos and fixed some naming conventions
