+++ manager/service/auth/claims.go
@@ -1,4 +1,4 @@
-package authservice
+package auth
 
 import "github.com/golang-jwt/jwt/v5"
