+++ manager/service/user/service.go
@@ -1,4 +1,4 @@
-package userservice
+package user
 
 import "github.com/ormushq/ormus/manager/entity"
